refactor: migrate func.lua to api.lua and add backends module- Replace  with  across all files- Add new backends module for handling streaming handlers- Update streaming logic to use backends module
