+++ lua/llm/app.lua
@@ -1,7 +1,7 @@
 local M = {}
 
 local conf = require("llm.config")
-local F = require("llm.common.func")
+local F = require("llm.common.api")
 local streaming = require("llm.common.streaming").GetStreamingOutput
 
 local state = require("llm.state")
