+++ lua/llm/common/tools.lua
@@ -6,7 +6,7 @@ local Split = require("nui.split")
 local Layout = require("nui.layout")
 local NuiText = require("nui.text")
 local conf = require("llm.config")
-local F = require("llm.common.func")
+local F = require("llm.common.api")
 local seamless = require("llm.common.seamless_border")
 local diff = require("llm.common.diff_style")
 local LOG = require("llm.common.log")
